Switch datatype yet again. %zu is just wrong and %llu is >python 2.7. sigh

+++ _pylibmcmodule.c
@@ -1689,7 +1689,7 @@ static PyObject *PylibMC_Client_set_behaviors(PylibMC_Client *self,
         if (r != MEMCACHED_SUCCESS) {
             PyErr_Format(PylibMCExc_MemcachedError,
                          "memcached_behavior_set returned %d for "
-                         "behavior '%.32s' = %zu", r, b->name, v);
+                         "behavior '%.32s' = %u", r, b->name, (unsigned int)v);
             goto error;
         }
     }
